refactor structure of organisms, tooltip updates, add compare section
