+++ src/components/organisms/HelpPanel/index.js
@@ -1,0 +1,1 @@
+/** scaffolding placeholder for help panel */
