+++ src/components/organisms/LocationPanel/LocationList.js
@@ -5,7 +5,6 @@ import { Typography } from '@material-ui/core';
 import LocationItem from './LocationItem';
 
 const LocationList = ({
-  summary, 
   metrics = ['avg', 'grd', 'coh'],
   demographic = 'all', 
   feature, className, 
@@ -13,15 +12,12 @@ const LocationList = ({
 }) => {
   return feature ? (
     <div className={classNames('location-list', className)}>
-      { summary && 
-        <Typography paragraph={true}>{summary}</Typography> 
-      }
       {
         others.map((f,i) =>
           f.properties.id !== feature.properties.id && 
             <LocationItem 
               key={'f'+i} 
-              number={i+1} 
+              idx={i} 
               feature={f}
               demographic={demographic}
               metrics={metrics}
