+++ src/components/organisms/LocationPanel/index.js
@@ -1,0 +1,3 @@
+export { default } from './LocationPanel';
+export { default as LocationPreview } from './LocationPreview';
+export { default as LocationStack } from './LocationStack';
