+++ src/components/organisms/Map/MapLegend.js
@@ -1,12 +1,12 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import classNames from 'classnames';
-import { getChoroplethColors, getMetricRange } from '../../modules/config';
-import LegendBar from '../molecules/LegendBar';
-import { getLang } from '../../modules/lang';
-import { getFeatureProperty } from '../../modules/features';
-import SedaScatterplotPreview from '../seda/SedaScatterplotPreview';
-import SedaLocationMarkers from '../seda/SedaLocationMarkers';
+import { getChoroplethColors, getMetricRange } from '../../../modules/config';
+import LegendBar from '../../molecules/LegendBar';
+import { getLang } from '../../../modules/lang';
+import { getFeatureProperty } from '../../../modules/features';
+import SedaScatterplotPreview from '../../seda/SedaScatterplotPreview';
+import SedaLocationMarkers from '../../seda/SedaLocationMarkers';
 import { Button } from '@material-ui/core';
 
 const choroplethColors = getChoroplethColors();
