+++ src/components/organisms/Map/MapVisualLegend.js
