+++ src/components/organisms/Map/constants.js
@@ -1,0 +1,8 @@
+/**
+ * default viewport on map view
+ */
+export const DEFAULT_VIEWPORT = {
+  latitude: 37.39,
+  longitude: -96.78,
+  zoom: 3.15
+}
