+++ src/components/organisms/Map/index.js
@@ -1,0 +1,3 @@
+export { default } from './MapBase';
+export { default as MapLegend } from './MapLegend';
+export { default as MapVisualLegend } from './MapVisualLegend';
