+++ src/components/organisms/Scatterplot/ScatterplotPreview.js
@@ -1,9 +1,10 @@
-import React, { useMemo, useEffect } from 'react'
+import React, { useMemo } from 'react'
 import PropTypes from 'prop-types'
-import SedaScatterplot, { fetchScatterplotVars } from 'react-seda-scatterplot'
-import { theme } from '../../style/echartTheme';
-import { getBaseVars } from '../../modules/config'
-import { getScatterplotOptions } from '../../style/scatterplot-style';
+import SedaScatterplot from 'react-seda-scatterplot'
+import { theme } from './echartTheme';
+import { getBaseVars } from '../../../modules/config'
+import { getScatterplotOptions } from './utils';
+import { getStateFipsFromAbbr } from '../../../constants/statesFips';
 
 const baseVars = getBaseVars()
 const endpoint = process.env.REACT_APP_VARS_ENDPOINT;
@@ -26,34 +27,27 @@ const getStateIds = (ids, fips) => {
  * @param {string} stateId 
  * @param {object} data 
  */
-const getStateHighlights = (stateId, data) => {
+const getStateHighlights = (stateId, data) => {  
   return data && data['name'] && stateId ? 
-    getStateIds(Object.keys(data['name']), stateId) : []
+    getStateIds(Object.keys(data['name']), getStateFipsFromAbbr(stateId)) : []
 }
 
-
-function DynamicScatterplot({
+function ScatterplotPreview({
   data,
   xVar,
   yVar,
   zVar,
   region,
   highlightedState,
-  variant,
   freeze,
-  error,
   children,
-  onHover,
-  onClick,
-  onData,
-  onReady,
   onError
 }) {
   // memoize scatterplot options
   const scatterplotOptions = useMemo(
     () => {
       return getScatterplotOptions(
-        variant, 
+        'preview', 
         data[region], 
         { xVar, yVar, zVar }, 
         highlightedState,
@@ -66,37 +60,15 @@ function DynamicScatterplot({
   const highlighted = useMemo(
     () => getStateHighlights(highlightedState, data && data[region]),
     [highlightedState, region, data[region]]
-  );  
-  // fetch any additional school level data for highlighted states
-  useEffect(() => {
-    if (!freeze && region === 'schools' && highlightedState && highlightedState !== 'us') {
-      fetchScatterplotVars(
-        [ xVar, yVar, zVar ], 
-        'schools', 
-        endpoint, 
-        getBaseVars()['schools'],
-        highlightedState
-      ).then((data) => {
-        onData && onData(data, 'schools');
-        return data
-      })
-    }
-  }, [xVar, yVar, zVar, region, highlightedState, freeze])
+  );
   return (
-    <div className='dynamic-scatterplot'>
-      { error &&
-        <span className="notification notification--error">{ error }</span>
-      }
-
+    <div className='scatterplot-preview'>
       <SedaScatterplot
         {...{
           endpoint,
           xVar,
           yVar,
           zVar,
-          onReady,
-          onClick,
-          onData,
           onError,
           data,
           highlighted,
@@ -106,18 +78,13 @@ function DynamicScatterplot({
         prefix={region}
         options={scatterplotOptions}
         metaVars={baseVars}
-        onHover={(loc) => loc && loc.id ?
-          onHover({ id: loc.id, properties: loc }) :
-          onHover(null)
-        }
       />
       {children}
     </div>
   )
 }
 
-DynamicScatterplot.propTypes = {
-  heading: PropTypes.object,
+ScatterplotPreview.propTypes = {
   xVar: PropTypes.string,
   yVar: PropTypes.string,
   zVar: PropTypes.string,
@@ -126,16 +93,11 @@ DynamicScatterplot.propTypes = {
   highlightedState: PropTypes.string,
   selected: PropTypes.array,
   hovered: PropTypes.object,
-  variant: PropTypes.string,
   children: PropTypes.node,
-  onHover: PropTypes.func,
-  onClick: PropTypes.func,
-  onData: PropTypes.func,
-  onReady: PropTypes.func,
   onError: PropTypes.func,
   freeze: PropTypes.bool,
   error: PropTypes.string,
 }
 
-export default DynamicScatterplot
+export default ScatterplotPreview
