+++ src/components/organisms/Scatterplot/index.js
@@ -1,0 +1,3 @@
+export { default } from './Scatterplot';
+export { default as ScatterplotOverlay } from './ScatterplotOverlay';
+export { default as ScatterplotPreview } from './ScatterplotPreview';
