+++ src/components/seda/SedaChart.js
@@ -9,7 +9,7 @@ import { getFeatureProperty } from '../../modules/features';
 import { getLang, hasLangKey } from '../../modules/lang';
 import { loadLocation, onHoverFeature, onScatterplotData, onScatterplotLoaded, onScatterplotError } from "../../actions";
 import LegendBar from '../molecules/LegendBar';
-import DynamicScatterplot from '../organisms/DynamicScatterplot';
+import Scatterplot from '../organisms/Scatterplot';
 import { Typography } from '@material-ui/core';
 import SedaLocationMarkers from './SedaLocationMarkers';
 
@@ -83,7 +83,7 @@ const SedaExplorerChart = ({
   const hoveredPrimary = hoveredValue || hoveredValue === 0 ? 
     Math.round(hoveredValue*100)/100 : null;
   return (
-    <DynamicScatterplot {...{
+    <Scatterplot {...{
       ...scatterplot,
       region,
       data,
@@ -121,7 +121,7 @@ const SedaExplorerChart = ({
           <span className="label">{rightLabel}</span>
         </div>
       }
-    </DynamicScatterplot>
+    </Scatterplot>
   )
 }
