+++ src/css/molecules/tooltip.css
@@ -1,3 +1,15 @@
+.tooltip__wrapper {
+  position:fixed;
+  pointer-events: none;
+  width: 100%;
+  height: 100%;
+  top:0;
+  bottom:0;
+  left:0;
+  right:0;
+  z-index: 999;
+}
+
 .tooltip {
   position: absolute;
   top:0;
