+++ src/css/organisms/_index.css
@@ -16,3 +16,4 @@
 @import 'map-legend.css';
 @import 'scatterplot-preview.css';
 @import 'stats-summary.css';
+@import 'location-list.css';
