+++ src/css/organisms/location-list.css
@@ -1,0 +1,15 @@
+.location-list .location-item {
+  position:relative;
+  min-height: 32px;
+}
+
+.location-list .location-item .location-item__heading {
+  --font-size: var(--small-font-size);
+  --font: var(--body-font);
+  font-weight: bold;
+}
+
+.location-list .location-item .location-item__marker {
+  left:-2px;
+  transform: scale(1);
+}
