+++ src/css/organisms/location-preview.css
@@ -4,7 +4,7 @@
   align-items: center;
   justify-content: space-between;
   width: 100%;
-  padding: 8px 16px 8px 48px;
+  padding: 8px 16px;
   height: 100%;
   background: #fff;
   border-right: 1px solid #eee;
@@ -15,26 +15,28 @@
 .location-preview:nth-child(1) {
   box-shadow: inset 1px 0 0 #eee;
 }
+
 .location-preview .location-item {
+  max-width: calc(100% - 16px);
+}
+
+
+
+.location-item {
   margin-bottom: 0;
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  justify-content: center;
+  padding-left: 32px;
 }
-.location-preview .location-item .location-item__name {
-  font-size: var(--card-title-size);
-  font-family: var(--heading-font);
-  font-weight: normal;
-  letter-spacing: 0;
-  line-height: 24px;
-  white-space: nowrap;
-  text-overflow: ellipsis;
-  overflow: hidden;
-  max-width: 100%;
-  max-height: 24px;
-  text-transform: none;
-}
-.location-preview .location-item .location-item__marker {
+
+.location-item .location-item__marker {
   position:absolute;
   left: 14px;
-  top: 18px;
+  top: 0px;
+  bottom:0;
+  margin:auto;
   text-shadow: 0 0 2px #000;
   border: 2px solid #fff;
   box-shadow: 0 0 2px #000;
@@ -42,19 +44,34 @@
   height: 24px;
   color: #fff;
 }
-.location-preview .stats--horizontal .stats__item {
-  width: 50%;
-  min-width: 80px;
+
+.location-item .location-item__heading {
+  --font: var(--heading-font);
+  --font-size: var(--card-title-size);
+
+  display: block;
   text-align: left;
+  position: relative;
+  font-size: var(--font-size);
+  font-family: var(--font);
+  font-weight: normal;
+  letter-spacing: 0;
+  line-height: 16px;
+  white-space: nowrap;
+  text-overflow: ellipsis;
+  overflow: hidden;
+  max-width: 100%;
+  max-height: 24px;
+  text-transform: none;
 }
 
-.location-preview .stats__item-marker {
-  position:absolute;
-  bottom:0;
-  left:0;
-  right:auto;
-  width:100%;
-  height: 8px;
-  width: 8px;
-  display: none;
+
+.location-item .location-item__content {
+  display: block; 
+  font-family: var(--font);
+  text-transform: initial;
+  letter-spacing: 0;
+  text-align:left;
+  line-height: 16px;
+  font-size: 13px;
 }
